+++ MapboxNavigation/NavigationMapView.swift
@@ -339,7 +339,6 @@ open class NavigationMapView: MGLMapView, UIGestureRecognizerDelegate {
             let function: CAMediaTimingFunction? = animated ? CAMediaTimingFunction(name: .linear) : nil
             let point = userAnchorPoint
             let padding = UIEdgeInsets(top: point.y, left: point.x, bottom: bounds.height - point.y, right: bounds.width - point.x)
-            // Omit padding when https://github.com/mapbox/mapbox-gl-native/pull/14081 has landed
             setCamera(newCamera, withDuration: duration, animationTimingFunction: function, edgePadding: padding, completionHandler: nil)
             userCourseView?.center = userAnchorPoint
         } else {
@@ -426,7 +425,7 @@ open class NavigationMapView: MGLMapView, UIGestureRecognizerDelegate {
     */
     public static let defaultPadding: UIEdgeInsets = UIEdgeInsets(top: 10, left: 20, bottom: 10, right: 20)
     
-    @objc public func showcase(_ routes: [Route], padding: UIEdgeInsets = NavigationMapView.defaultPadding, animated: Bool = false) {
+    @objc public func showcase(_ routes: [Route], animated: Bool = false) {
         guard let active = routes.first,
               let coords = active.coordinates,
               !coords.isEmpty else { return } //empty array
@@ -438,15 +437,15 @@ open class NavigationMapView: MGLMapView, UIGestureRecognizerDelegate {
         showRoutes(routes)
         showWaypoints(active)
         
-        fit(to: active, facing: 0, padding: padding, animated: animated)
+        fit(to: active, facing: 0, animated: animated)
     }
     
-    func fit(to route: Route, facing direction:CLLocationDirection = 0, padding: UIEdgeInsets = NavigationMapView.defaultPadding, animated: Bool = false) {
+    func fit(to route: Route, facing direction:CLLocationDirection = 0, animated: Bool = false) {
         guard let coords = route.coordinates, !coords.isEmpty else { return }
       
-        setUserTrackingMode(.none, animated: false)
+        setUserTrackingMode(.none, animated: false, completionHandler: nil)
         let line = MGLPolyline(coordinates: coords, count: UInt(coords.count))
-        let camera = cameraThatFitsShape(line, direction: direction, edgePadding: padding)
+        let camera = cameraThatFitsShape(line, direction: direction, edgePadding: .zero)
         
         setCamera(camera, animated: animated)
     }
@@ -817,32 +816,36 @@ open class NavigationMapView: MGLMapView, UIGestureRecognizerDelegate {
         var linesPerLeg: [MGLPolylineFeature] = []
         
         for (index, leg) in route.legs.enumerated() {
+            let lines: [MGLPolylineFeature]
             // If there is no congestion, don't try and add it
-            guard let legCongestion = leg.segmentCongestionLevels, legCongestion.count < coordinates.count else {
-                return [MGLPolylineFeature(coordinates: route.coordinates!, count: UInt(route.coordinates!.count))]
-            }
-            
-            // The last coord of the preceding step, is shared with the first coord of the next step, we don't need both.
-            let legCoordinates: [CLLocationCoordinate2D] = leg.steps.enumerated().reduce([]) { allCoordinates, current in
-                let index = current.offset
-                let step = current.element
-                let stepCoordinates = step.coordinates!
-                
-                return index == 0 ? stepCoordinates : allCoordinates + stepCoordinates.suffix(from: 1)
-            }
-            
-            let mergedCongestionSegments = combine(legCoordinates, with: legCongestion)
-            
-            let lines: [MGLPolylineFeature] = mergedCongestionSegments.map { (congestionSegment: CongestionSegment) -> MGLPolylineFeature in
-                let polyline = MGLPolylineFeature(coordinates: congestionSegment.0, count: UInt(congestionSegment.0.count))
-                polyline.attributes[MBCongestionAttribute] = String(describing: congestionSegment.1)
-                polyline.attributes["isAlternateRoute"] = false
-                if let legIndex = legIndex {
-                    polyline.attributes[MBCurrentLegAttribute] = index == legIndex
-                } else {
-                    polyline.attributes[MBCurrentLegAttribute] = index == 0
+            if let legCongestion = leg.segmentCongestionLevels, legCongestion.count < coordinates.count {
+                // The last coord of the preceding step, is shared with the first coord of the next step, we don't need both.
+                let legCoordinates: [CLLocationCoordinate2D] = leg.steps.enumerated().reduce([]) { allCoordinates, current in
+                    let index = current.offset
+                    let step = current.element
+                    let stepCoordinates = step.coordinates!
+                    
+                    return index == 0 ? stepCoordinates : allCoordinates + stepCoordinates.suffix(from: 1)
                 }
-                return polyline
+                
+                let mergedCongestionSegments = combine(legCoordinates, with: legCongestion)
+                
+                lines = mergedCongestionSegments.map { (congestionSegment: CongestionSegment) -> MGLPolylineFeature in
+                    let polyline = MGLPolylineFeature(coordinates: congestionSegment.0, count: UInt(congestionSegment.0.count))
+                    polyline.attributes[MBCongestionAttribute] = String(describing: congestionSegment.1)
+                    return polyline
+                }
+            } else {
+                lines = [MGLPolylineFeature(coordinates: route.coordinates!, count: UInt(route.coordinates!.count))]
+            }
+            
+            for line in lines {
+                line.attributes["isAlternateRoute"] = false
+                if let legIndex = legIndex {
+                    line.attributes[MBCurrentLegAttribute] = index == legIndex
+                } else {
+                    line.attributes[MBCurrentLegAttribute] = index == 0
+                }
             }
             
             linesPerLeg.append(contentsOf: lines)